+++ source/extensions/omni.isaac.lab_tasks/docs/CHANGELOG.rst
@@ -1,13 +1,23 @@
 Changelog
 ---------
 
-0.7.10 (2024-07-02)
+0.8.0 (2024-07-26)
 ~~~~~~~~~~~~~~~~~~
+
+Removed
+^^^^^^^
+
+* Renamed the action term names inside the manager-based lift-manipulation task. Earlier, they were called
+  ``body_joint_pos`` and ``gripper_joint_pos``. Now, they are called ``arm_action`` and ``gripper_action``.
+
+
+0.7.10 (2024-07-02)
+~~~~~~~~~~~~~~~~~~~
 
 Added
 ^^^^^
 
-* Extended skrl wrapper to support training/evaluation using JAX
+* Extended skrl wrapper to support training/evaluation using JAX.
 
 
 0.7.9 (2024-07-01)